+++ src/main/java/com/ds/app/repository/iAppUserRepository.java
@@ -1,14 +1,13 @@
 package com.ds.app.repository;
 
-import java.util.Optional;
-
+import com.ds.app.entity.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ds.app.entity.AppUser;
+import java.util.Optional;
 
 @Repository
-public interface iAppUserRepository extends JpaRepository<AppUser,Integer>{
+public interface iAppUserRepository extends JpaRepository<AppUser,Long>{
 
 	public Optional<AppUser> findByUsername(String username);
 }
